fix: cant delete something that doesn't exist

+++ common/twilio_helpers.py
@@ -61,7 +61,8 @@ def if_exists_active_cm(user_channel=None, user_identity=None, service=None):
 def swap_cm_with_active(old_cm=None, channel=None):
     from common.helpers import active_cm_list
     active_cm = choice(active_cm_list)
-    old_cm.delete()
+    if old_cm:
+        old_cm.delete()
     channel.members.create(
         active_cm,
         attributes=json.dumps(active_cm_attributes)
